vps_worker/requirements.txt
# Telegram MTProto client
Telethon==1.34.0

# Database
supabase==2.4.0

# HTTP requests
aiohttp==3.9.3
requests==2.31.0

# Environment & Config
python-dotenv==1.0.1

# Async utilities
asyncio-throttle==1.0.2

# Scheduling
APScheduler==3.10.4

# Timezone
pytz==2024.1

# Proxy support
PySocks==1.7.1

# AI/LLM for content generation (optional)
openai==1.12.0

# SMS services
# onlinesim-python==0.1.0  # Optional, we use direct API

# Logging & Monitoring
structlog==24.1.0

# Security
cryptography==42.0.2

# Rate limiting
ratelimit==2.2.1

# Retry logic
tenacity==8.2.3

# CLI colors
colorama==0.4.6

# Validation
pydantic==2.6.1

# Process management
supervisor==4.2.5
